LinenoiseRemotePrint: remotely pass text to print
When other processes called printf and linenoise had a prompt displayed,
the text got printed after the prompt and looked horrible. This function
restarts the line before printing with a \r, prints, then pushes a
ctrl-r to a special pipe that gets waited on alongside stdin. That
ctrl-r now triggers the refresh line function. The result is your text
is printed, then the prompt + incomplete input is redrawn below.